updated all the scripts

+++ Ec2_Create.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 
-#COMPONENT=$1
+APP=$1
 
-#if [ -z "$1" ]
-#  then
-#  echo -e " component name as an argument is mandatory \t\t"
-#  exit 1
-#fi
+if [ -z "$1" ]
+  then
+  echo -e " provide APP name as command line arguement \t\t"
+  exit 1
+fi
 
 rm -rf akey*
 
@@ -20,4 +20,4 @@ echo "Launching AWS ec2 Instance.."
 
 #SYNTAX: aws ec2 run-instances [other options] --tag-specifications 'ResourceType=instance,Tags=[{Key=Name,Value=MyInstanceName}]'
 
-aws ec2 run-instances --image-id $AMI_ID --instance-type t2.micro --key-name akey  --tag-specifications 'ResourceType=instance,Tags=[{Key=Name,Value=cliserver}]'|jq
+aws ec2 run-instances --image-id $AMI_ID --instance-type t2.micro --key-name akey  --tag-specifications 'ResourceType=instance,Tags=[{Key=Name,Value=$APP}]'|jq
